Add api/document/from-url end point to add new documents from a url and used it directly from client side. Also, made a little update in error handling of indra connecter.
